Fix conflict with Projax/Prototype used by Mantis
There was a conflict with the Projax/Prototype libraries included with
MantisBT over the usage of the $() construct.  To solve this, jQuery
will only listen on $() inside of jQuery(document).ready().

This means that any plugins using $() outside of .ready() will need to
update their scripts to follow the new expectations.